Renaming files, refactoring tests
