+++ src/groovy/co/tpaga/grails/plugin/springsecurity/rest/RestApiKeyAuthenticationSuccessHandler.groovy
