+++ src/groovy/co/tpaga/grails/plugin/springsecurity/rest/apiKey/basic/HTTPBasicAuthApiKeyAuthenticationFailureHandler.groovy
@@ -1,6 +1,7 @@
-package co.tpaga.grails.plugin.springsecurity.rest.token.bearer
+package co.tpaga.grails.plugin.springsecurity.rest.apiKey.basic
 
-import co.tpaga.grails.plugin.springsecurity.rest.token.storage.ApiKeyNotFoundException
+import co.tpaga.grails.plugin.springsecurity.rest.apiKey.storage.ApiKeyNotFoundException
+import groovy.util.logging.Slf4j
 import org.springframework.security.core.AuthenticationException
 import org.springframework.security.web.authentication.AuthenticationFailureHandler
 
@@ -8,9 +9,11 @@ import javax.servlet.ServletException
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
-class HTTPBasicBearerApiKeyAuthenticationFailureHandler implements AuthenticationFailureHandler {
+@Slf4j
+class HTTPBasicAuthApiKeyAuthenticationFailureHandler implements AuthenticationFailureHandler {
 
     String realm
+    Integer statusCode
 
     /**
      * Sends the proper response code and headers, as defined by RFC26147.
@@ -34,9 +37,7 @@ class HTTPBasicBearerApiKeyAuthenticationFailureHandler implements Authenticatio
             }
         }
 
-        if (response.status == 200) {
-            response.status = 401
-        }
+        response.status = statusCode
 
         log.debug "Sending status code ${response.status} and header WWW-Authenticate: ${response.getHeader('WWW-Authenticate')}"
     }
