+++ src/groovy/co/tpaga/grails/plugin/springsecurity/rest/apiKey/reader/ApiKeyReader.groovy
@@ -1,4 +1,4 @@
-package co.tpaga.grails.plugin.springsecurity.rest.token.reader
+package co.tpaga.grails.plugin.springsecurity.rest.apiKey.reader
 
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
@@ -6,11 +6,11 @@ import javax.servlet.http.HttpServletResponse
 public interface ApiKeyReader {
 
     /**
-     * Reads a token (if any) from the request
+     * Reads an Api Key (if any) from the request
      *
      * @param request the HTTP request
      * @param response the response, in case any status code has to be sent
-     * @return the token when found, null otherwise
+     * @return the Api Key when found, null otherwise
      */
     String findApiKey(HttpServletRequest request, HttpServletResponse response)
