+++ test/unit/co/tpaga/grails/plugin/springsecurity/rest/HTTPBasicApiKeyReaderSpec.groovy
@@ -1,6 +1,6 @@
 package co.tpaga.grails.plugin.springsecurity.rest
 
-import co.tpaga.grails.plugin.springsecurity.rest.token.reader.HTTPBasicApiKeyReader
+import co.tpaga.grails.plugin.springsecurity.rest.apiKey.reader.HTTPBasicApiKeyReader
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.mock.web.MockHttpServletResponse
 import spock.lang.Specification
@@ -17,7 +17,7 @@ class HTTPBasicApiKeyReaderSpec extends Specification {
 
 
     def "api key value can be read from Authorization header in a #method request"() {
-        def token  = 'my_api_key'
+        def apiKey  = 'my_api_key'
 
         //base64.encode(my_api_key:) includes :
 
@@ -25,7 +25,7 @@ class HTTPBasicApiKeyReaderSpec extends Specification {
         request.method = method
 
         expect:
-        apiKeyReader.findApiKey(request, response) == token
+        apiKeyReader.findApiKey(request, response) == apiKey
 
         where:
         method << [ 'GET', 'POST', 'PUT', 'PATCH', 'DELETE', 'HEAD', 'OPTIONS' ]
